Update v3.0 build 7
* change backup process
* add backup-time and max-backups to config

+++ myrpg/src/org/caliog/myRPG/Entities/PlayerManager.java
@@ -107,7 +107,7 @@ public class PlayerManager {
 	return null;
     }
 
-    public static void task(float seconds) {
+    public static void task(long seconds) {
 	for (myClass clazz : players.values()) {
 	    int i = clazz.getIntelligence();
 	    int s = 1;
@@ -125,7 +125,7 @@ public class PlayerManager {
 		clazz.regainFood();
 	    }
 	    if (clazz.getHealth() != 0.0D) {
-		if (seconds % 5.0F == 0.0F) {
+		if (seconds % 5 == 0.0F) {
 		    clazz.addHealth(clazz.getMaxHealth() * 0.05D);
 		}
 		double d = clazz.getHealth() / clazz.getMaxHealth();
